restructure modules python module
The following was the reason for this change:
From other places in mythril there is only a dependency towards base,
and a future loader. By structuring the module in this way, this
lack of dependency is made more explicit.
