+++ mythril/analysis/module/__init__.py
@@ -1,0 +1,2 @@
+from mythril.analysis.module.base import EntryPoint, DetectionModule
+from mythril.analysis.module.loader import ModuleLoader
