+++ mythril/analysis/module/modules/arbitrary_write.py
@@ -1,6 +1,6 @@
 """This module contains the detection code for arbitrary storage write."""
 import logging
-from mythril.analysis.modules.base import DetectionModule, EntryPoint
+from mythril.analysis.module.base import DetectionModule, EntryPoint
 from mythril.analysis.potential_issues import (
     get_potential_issues_annotation,
     PotentialIssue,
